H.Res. 86 Engrossed in House
Providing for consideration of the resolution (H. Res. 78) amending the Rules of the House of Representatives to permit Delegates and the Resident Commissioner to the Congress to cast votes in the Committee of the Whole House on the state of the Union.

Version:  Engrossed in House (eh)
Date:     2007-01-24
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hres86eh/xml/BILLS-110hres86eh.xml

Sponsored-By: H000324
Roll-Call: House 110-1-51 2007-01-24 Passed 229-191
Roll-Call: House 110-1-52 2007-01-24 Passed 229-189
Roll-Call: House 110-1-53 2007-01-24 Passed 228-188