H.R. 2919 Introduced in House
To authorize the Secretary of the Interior to conduct a study to determine the suitability and feasibility of designating the Virgin Islands Military and Veterans Memorial, to be located in Fredericksted, St. Croix, U.S. Virgin Islands, as a unit of the National Park System.

Version:  Introduced in House (ih)
Date:     2007-06-28
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hr2919ih/xml/BILLS-110hr2919ih.xml

Sponsored-By: C000380