H.Res. 559 Introduced in House
Providing for consideration of the resolution (H. Res. 479) to amend the Rules of the House of Representatives to provide for enforcement of clause 9 of rule XXI of the Rules of the House of Representatives.

Version:  Introduced in House (ih)
Date:     2007-07-19
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hres559ih/xml/BILLS-110hres559ih.xml

Sponsored-By: B000589