H.R. 2120 Reported in House
To direct the Secretary of the Interior to proclaim as reservation for the benefit of the Sault Ste. Marie Tribe of Chippewa Indians a parcel of land now held in trust by the United States for that Indian tribe.

Version:  Reported in House (rh)
Date:     2007-07-30
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hr2120rh/xml/BILLS-110hr2120rh.xml

Sponsored-By: S001045