H.R. 2120 Referred in Senate
To direct the Secretary of the Interior to proclaim as reservation for the benefit of the Sault Ste. Marie Tribe of Chippewa Indians a parcel of land now held in trust by the United States for that Indian tribe.

Version:  Referred in Senate (rfs)
Date:     2007-08-03
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hr2120rfs/xml/BILLS-110hr2120rfs.xml

Sponsored-By: S001045