H.Res. 893 Engrossed in House
Providing for the consideration of the Senate amendment to the House amendment to the Senate amendment to the bill (H.R. 2764) making appropriations for the Department of State, foreign operations, and related programs for the fiscal year ending September 30, 2008, and for consideration of the joint resolution (H.J. Res. 72) making further continuing appropriations for the fiscal year 2008, and for other purposes.

Version:  Engrossed in House (eh)
Date:     2007-12-19
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hres893eh/xml/BILLS-110hres893eh.xml

Sponsored-By: M000312