S. 2590 Introduced in Senate
A bill to authorize the Secretary of the Interior, acting through the Director of the National Park Service, to designate the Dr. Norman E. Borlaug Birthplace and Childhood Home in Cresco, Iowa, as a National Historic Site and as a unit of the National Park System, and for other purposes.

Version:  Introduced in Senate (is)
Date:     2008-02-04
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110s2590is/xml/BILLS-110s2590is.xml

Sponsored-By: G000386
Cosponsor-Count: 1