H.Res. 982 Reported in House
Providing for the adoption of the resolution (H. Res. 979) recommending that the House of Representatives find Harriet Miers and Joshua Bolten, Chief of Staff, White House, in contempt of Congress for refusal to comply with subpoenas duly issued by the Committee on the Judiciary and for the adoption of the resolution (H. Res. 980) authorizing the Committee on the Judiciary to initiate or intervene in judicial proceedings to enforce certain subpoenas.

Version:  Reported in House (rh)
Date:     2008-02-13
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hres982rh/xml/BILLS-110hres982rh.xml

Sponsored-By: S000480