H.Res. 1146 Introduced in House
Expressing the sense of the House of Representatives that the International Joint Commission should adopt a water level management plan for Lake Ontario and the St. Lawrence River that strongly takes into account environmental considerations and the concerns of the public and the affected States and maximizes hydropower production at existing facilities, and further urges the Secretary of State not to approve a plan that fails to do so.

Version:  Introduced in House (ih)
Date:     2008-04-24
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hres1146ih/xml/BILLS-110hres1146ih.xml

Sponsored-By: M000472
Cosponsor-Count: 1