S.Res. 626 Introduced in Senate
A resolution expressing the sense of the Senate that the Supreme Court of the United States erroneously decided Kennedy v. Louisiana, No. 07-343 (2008), and that the eighth amendment to the Constitution of the United States allows the imposition of the death penalty for the rape of a child.

Version:  Introduced in Senate (is)
Date:     2008-07-25
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110sres626is/xml/BILLS-110sres626is.xml

Sponsored-By: V000127
Cosponsor-Count: 8