H.Res. 1434 Reported in House
Providing for consideration of the bill (H.R. 6842) to require the District of Columbia to revise its laws regarding the use and possession of firearms as necessary to comply with the requirements of the decision of the Supreme Court in the case of District of Columbia v. Heller, in a manner that protects the security interests of the Federal government and the people who work in, reside in, or visit the District of Columbia and does not undermine the efforts of law enforcement, homeland security, and military officials to protect the Nation's capital from crime and terrorism.

Version:  Reported in House (rh)
Date:     2008-09-15
Congress: 110

Source: https://www.govinfo.gov/content/pkg/BILLS-110hres1434rh/xml/BILLS-110hres1434rh.xml

Sponsored-By: M000312